Revert "Feat/f001 (#1)"
This reverts commit 3d8dfd46eade933ac7a6857417f821fd502f55c4.
