+++ .gitignore
@@ -31,6 +31,3 @@ build/
 
 ### VS Code ###
 .vscode/
-
-### Custom ###
-HolidaySwap.iml
