+++ src/main/java/com/example/holidayswap/Application.java
@@ -2,11 +2,9 @@ package com.example.holidayswap;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@EnableCaching
 public class Application {
 
     public static void main(String[] args) {
