+++ src/main/java/com/example/holidayswap/config/security/JwtAuthenticationFilter.java
@@ -33,42 +33,36 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
             @NonNull HttpServletResponse response,
             @NonNull FilterChain filterChain
     ) throws ServletException, IOException {
-        try {
-            if (request.getServletPath().contains("/api/v1/auth")) {
-                filterChain.doFilter(request, response);
-                return;
-            }
-            final String authHeader = request.getHeader("Authorization");
-            final String jwt;
-            final String userEmail;
-            if (authHeader == null || !authHeader.startsWith("Bearer ")) {
-                filterChain.doFilter(request, response);
-                return;
-            }
-            jwt = authHeader.substring(7);
-            userEmail = jwtService.extractUsername(jwt);
-            if (userEmail != null && SecurityContextHolder.getContext().getAuthentication() == null) {
-                UserDetails userDetails = this.userDetailsService.loadUserByUsername(userEmail);
-                var isTokenValid = tokenRepository.findByValueEquals(jwt)
-                        .map(t -> t.getStatus().equals(TokenStatus.VALID) && t.getExpirationTime().isAfter(java.time.LocalDateTime.now()))
-                        .orElse(false);
-                if (jwtService.isTokenValid(jwt, (User) userDetails) && Boolean.TRUE.equals(isTokenValid)) {
-                    UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
-                            userDetails,
-                            null,
-                            userDetails.getAuthorities()
-                    );
-                    authToken.setDetails(
-                            new WebAuthenticationDetailsSource().buildDetails(request)
-                    );
-                    SecurityContextHolder.getContext().setAuthentication(authToken);
-                }
-            }
+        if (request.getServletPath().contains("/api/v1/auth")) {
             filterChain.doFilter(request, response);
-        } catch (Exception e) {
-            logger.error("Error in JwtAuthenticationFilter, cannot set user authentication: {}", e);
-            response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Error in JwtAuthenticationFilter, cannot set user authentication");
-            filterChain.doFilter(request, response);
+            return;
         }
+        final String authHeader = request.getHeader("Authorization");
+        final String jwt;
+        final String userEmail;
+        if (authHeader == null ||!authHeader.startsWith("Bearer ")) {
+            filterChain.doFilter(request, response);
+            return;
+        }
+        jwt = authHeader.substring(7);
+        userEmail = jwtService.extractUsername(jwt);
+        if (userEmail != null && SecurityContextHolder.getContext().getAuthentication() == null) {
+            UserDetails userDetails = this.userDetailsService.loadUserByUsername(userEmail);
+            var isTokenValid = tokenRepository.findByValueEquals(jwt)
+                    .map(t -> t.getStatus().equals(TokenStatus.VALID) && t.getExpirationTime().isAfter(java.time.LocalDateTime.now()))
+                    .orElse(false);
+            if (jwtService.isTokenValid(jwt, (User) userDetails) && isTokenValid) {
+                UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
+                        userDetails,
+                        null,
+                        userDetails.getAuthorities()
+                );
+                authToken.setDetails(
+                        new WebAuthenticationDetailsSource().buildDetails(request)
+                );
+                SecurityContextHolder.getContext().setAuthentication(authToken);
+            }
+        }
+        filterChain.doFilter(request, response);
     }
 }
