--- src/main/java/com/example/holidayswap/constants/ErrorMessage.java
@@ -1,18 +1,0 @@
-package com.example.holidayswap.constants;
-
-public class ErrorMessage {
-    public static final String JWT_TOKEN_INVALID = "JWT token is expired or invalid.";
-    public static final String VERIFICATION_TOKEN_INVALID = "Verification token invalid.";
-    public static final String PASSWORD_RESET_TOKEN_INVALID = "Password reset token is expired or invalid.";
-    public static final String TOKEN_EXPIRED = "Token expired.";
-    public static final String TOKEN_REVOKED = "Token already used.";
-    public static final String USER_NOT_FOUND = "User not found";
-    public static final String USER_NOT_VERIFIED = "User not verified.";
-    public static final String USER_BLOCKED = "User is blocked.";
-    public static final String EMAIL_NOT_FOUND = "Email not found.";
-    public static final String EMAIL_HAS_ALREADY_BEEN_TAKEN = "Email has already been taken.";
-    public static final String ROLE_NOT_FOUND = "Role not found.";
-    public static final String ROLE_ALREADY_EXISTS = "Role already exists.";
-    public static final String PROFILE_NOT_FOUND = "Profile not found.";
-    public static final String INCORRECT_EMAIL_OR_PASSWORD = "Incorrect email or password.";
-}
