+++ src/main/java/com/example/holidayswap/controller/auth/UserController.java
@@ -1,14 +1,13 @@
 package com.example.holidayswap.controller.auth;
 
-import com.example.holidayswap.domain.dto.request.auth.RegisterRequest;
-import com.example.holidayswap.domain.dto.request.auth.UserRequest;
 import com.example.holidayswap.domain.dto.response.auth.UserProfileResponse;
 import com.example.holidayswap.service.auth.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
@@ -19,33 +18,5 @@ public class UserController {
     @GetMapping("/{userId}")
     public ResponseEntity<UserProfileResponse> getUserById(@PathVariable("userId") Long userId) {
         return ResponseEntity.ok(userService.getUserById(userId));
-    }
-
-    @GetMapping("/profile")
-    public ResponseEntity<UserProfileResponse> getUserInfo() {
-        return ResponseEntity.ok(userService.getUserInfo());
-    }
-
-    @GetMapping("/search")
-    public List<UserProfileResponse> findAllByEmailNamePhoneWithPagination(
-            @RequestParam(name = "email", defaultValue = "") String email,
-            @RequestParam(name = "name", defaultValue = "") String name,
-            @RequestParam(name = "phone", defaultValue = "") String phone,
-            @RequestParam(name = "limit", defaultValue = "20") Integer limit,
-            @RequestParam(name = "offset", defaultValue = "0") Integer offset
-    ) {
-        return userService.findAllByEmailNamePhoneWithPagination(email, name, phone, limit, offset);
-    }
-
-    @DeleteMapping("/{userId}")
-    public ResponseEntity<Void> deleteUser(@PathVariable("userId") Long userId) {
-        userService.deleteUser(userId);
-        return ResponseEntity.noContent().build();
-    }
-
-    @PostMapping
-    public ResponseEntity<Void> registerUser(@RequestBody UserRequest userRequest) {
-        userService.createUser(userRequest);
-        return ResponseEntity.noContent().build();
     }
 }
