--- src/main/java/com/example/holidayswap/controller/chat/WebSocketController.java
@@ -1,17 +1,0 @@
-package com.example.holidayswap.controller.chat;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/websocket")
-public class WebSocketController {
-    private final SimpMessagingTemplate messagingTemplate;
-
-    @PostMapping
-    public void send(@RequestParam("destination") String destination, @RequestBody Object request) {
-        messagingTemplate.convertAndSend(destination, request);
-    }
-}
