+++ src/main/java/com/example/holidayswap/domain/dto/request/auth/LoginRequest.java
@@ -1,12 +1,8 @@
 package com.example.holidayswap.domain.dto.request.auth;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Data
-@NoArgsConstructor
-@AllArgsConstructor
 public class LoginRequest {
     private String email;
     private String password;
