+++ src/main/java/com/example/holidayswap/domain/entity/auth/User.java
@@ -5,14 +5,15 @@ import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.hibernate.annotations.Nationalized;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import java.io.Serial;
-import java.io.Serializable;
 import java.time.LocalDate;
 import java.util.Collection;
 import java.util.Collections;
@@ -23,11 +24,8 @@ import java.util.List;
 @AllArgsConstructor
 @Builder
 @Entity
-@ToString
 @Table(name = "users")
-public class User implements UserDetails, Serializable {
-    @Serial
-    private static final long serialVersionUID = 1L;
+public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(
@@ -44,7 +42,7 @@ public class User implements UserDetails, Serializable {
     private String email;
 
     @NotBlank(message = "Password must be specified.")
-    @Column(nullable = false, name = "password_hash")
+    @Column(nullable = false)
     private String passwordHash;
 
     @NotEmpty(message = "Username must be specified.")
@@ -53,10 +51,6 @@ public class User implements UserDetails, Serializable {
             unique = true
     )
     private String username;
-
-    @Nationalized
-    @Column(name = "full_name")
-    private String fullName;
 
     @Column
     private String avatar;
