+++ src/main/java/com/example/holidayswap/domain/mapper/auth/UserMapper.java
@@ -1,7 +1,6 @@
 package com.example.holidayswap.domain.mapper.auth;
 
 import com.example.holidayswap.domain.dto.request.auth.RegisterRequest;
-import com.example.holidayswap.domain.dto.request.auth.UserRequest;
 import com.example.holidayswap.domain.dto.response.auth.UserProfileResponse;
 import com.example.holidayswap.domain.entity.auth.User;
 import org.mapstruct.Mapper;
@@ -16,7 +15,4 @@ public interface UserMapper {
 
     @Mapping(source = "password", target = "passwordHash", ignore = true)
     User toUserEntity(RegisterRequest registerRequest);
-
-    @Mapping(source = "password", target = "passwordHash", ignore = true)
-    User toUserEntity(UserRequest userRequest);
 }
