--- src/main/java/com/example/holidayswap/service/auth/AuthenticationServiceImpl.java
@@ -1,249 +1,0 @@
-package com.example.holidayswap.service.auth;
-
-import com.example.holidayswap.config.security.JwtService;
-import com.example.holidayswap.domain.dto.request.auth.LoginRequest;
-import com.example.holidayswap.domain.dto.request.auth.RegisterRequest;
-import com.example.holidayswap.domain.dto.request.auth.ResetPasswordRequest;
-import com.example.holidayswap.domain.dto.response.auth.AuthenticationResponse;
-import com.example.holidayswap.domain.entity.auth.*;
-import com.example.holidayswap.domain.exception.AccessDeniedException;
-import com.example.holidayswap.domain.exception.DataIntegrityViolationException;
-import com.example.holidayswap.domain.exception.EntityNotFoundException;
-import com.example.holidayswap.domain.exception.VerificationException;
-import com.example.holidayswap.domain.mapper.auth.UserMapper;
-import com.example.holidayswap.repository.auth.TokenRepository;
-import com.example.holidayswap.repository.auth.UserRepository;
-import com.example.holidayswap.service.EmailService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.concurrent.CompletableFuture;
-
-import static com.example.holidayswap.constants.ErrorMessage.*;
-
-@Service
-@Slf4j
-@RequiredArgsConstructor
-public class AuthenticationServiceImpl implements AuthenticationService {
-    private final UserRepository userRepository;
-    private final TokenRepository tokenRepository;
-    private final EmailService emailService;
-    private final JwtService jwtService;
-    private final PasswordEncoder passwordEncoder;
-    private final AuthenticationManager authenticationManager;
-    @Value("${application.security.jwt.expiration}")
-    private long jwtExpiration;
-    @Value("${application.security.jwt.access-token.expiration}")
-    private long accessTokenExpiration;
-    @Value("${application.security.jwt.refresh-token.expiration}")
-    private long refreshExpiration;
-
-    @Override
-    @Transactional
-    public void register(RegisterRequest request) {
-        userRepository.getUserByEmailEquals(request.getEmail())
-                .ifPresent(user -> {
-                    throw new DataIntegrityViolationException(EMAIL_HAS_ALREADY_BEEN_TAKEN);
-                });
-        var user = UserMapper.INSTANCE.toUserEntity(request);
-        user.setStatus(UserStatus.PENDING);
-        user.setPasswordHash(passwordEncoder.encode(request.getPassword()));
-        userRepository.save(user);
-        try {
-            emailService.sendRegistrationReceipt(user.getEmail(), user.getUsername());
-            var token = tokenRepository.save(Token
-                    .builder()
-                    .user(user)
-                    .tokenType(TokenType.EMAIL_VERIFICATION)
-                    .expirationTime(LocalDateTime.now().plusSeconds(jwtExpiration / 1000))
-                    .status(TokenStatus.VALID)
-                    .value(jwtService.generateToken(new HashMap<>(), user))
-                    .build());
-            emailService.sendVerificationEmail(user.getEmail(), token.getValue());
-        } catch (Exception e) {
-            log.error("Error sending verification email", e);
-        }
-    }
-
-    @Override
-    public AuthenticationResponse login(LoginRequest loginRequest) {
-        authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(
-                        loginRequest.getEmail(),
-                        loginRequest.getPassword()
-                )
-        );
-        var user = userRepository.getUserByEmailEquals(loginRequest.getEmail())
-                .orElseThrow(() -> new EntityNotFoundException(EMAIL_NOT_FOUND));
-        if (user.getStatus().equals(UserStatus.PENDING)) {
-            throw new AccessDeniedException(USER_NOT_VERIFIED);
-        }
-        if (user.getStatus().equals(UserStatus.BLOCKED)) {
-            throw new AccessDeniedException(USER_BLOCKED);
-        }
-        return getAuthenticationResponse(user);
-    }
-
-    private void saveUserToken(User user, String jwtToken, TokenType tokenType, Long expirationTime) {
-        var token = Token.builder()
-                .user(user)
-                .value(jwtToken)
-                .tokenType(tokenType)
-                .status(TokenStatus.VALID)
-                .expirationTime(LocalDateTime.now().plusSeconds(expirationTime / 1000))
-                .build();
-        tokenRepository.save(token);
-    }
-
-    @Override
-    public void confirmRegistration(Long userId) {
-        userRepository.findById(userId)
-                .ifPresentOrElse(u -> {
-                    u.setStatus(UserStatus.ACTIVE);
-                    userRepository.save(u);
-                }, () -> {
-                    throw new EntityNotFoundException(USER_NOT_FOUND);
-                });
-    }
-
-    @Override
-    public void forgetPassword(String email) {
-        var user = userRepository.getUserByEmailEquals(email)
-                .orElseThrow(() -> new EntityNotFoundException(USER_NOT_FOUND));
-        var token = tokenRepository.save(Token
-                .builder()
-                .user(user)
-                .tokenType(TokenType.EMAIL_VERIFICATION)
-                .expirationTime(LocalDateTime.now().plusSeconds(jwtExpiration / 1000))
-                .status(TokenStatus.VALID)
-                .value(jwtService.generateToken(new HashMap<>(), user))
-                .build());
-        emailService.sendForgotPasswordEmail(user.getEmail(), token.getValue());
-    }
-
-    @Override
-    public void resetPassword(ResetPasswordRequest resetPasswordRequest) {
-        String userEmail = jwtService.extractUsername(resetPasswordRequest.getToken());
-        if (userEmail == null) {
-            throw new VerificationException(USER_NOT_FOUND);
-        }
-        var user = this.userRepository.getUserByEmailEquals(userEmail)
-                .orElseThrow(() -> new EntityNotFoundException(USER_NOT_FOUND));
-        if (!jwtService.isTokenValid(resetPasswordRequest.getToken(), user)) {
-            throw new VerificationException(PASSWORD_RESET_TOKEN_INVALID);
-        }
-        tokenRepository.findByValueEquals(resetPasswordRequest.getToken())
-                .ifPresent(token -> {
-                    if (token.getStatus().equals(TokenStatus.VALID)) {
-                        token.setStatus(TokenStatus.REVOKED);
-                        tokenRepository.save(token);
-                        user.setPasswordHash(passwordEncoder.encode(resetPasswordRequest.getPassword()));
-                        userRepository.save(user);
-                        CompletableFuture.runAsync(() -> {
-                            var validUserPasswordResetTokens = tokenRepository.findAllValidTokenByUser(user.getUserId())
-                                    .stream()
-                                    .filter(t -> t.getTokenType().equals(TokenType.PASSWORD_RESET))
-                                    .map(t -> {
-                                        t.setStatus(TokenStatus.REVOKED);
-                                        return t;
-                                    })
-                                    .toList();
-                            tokenRepository.saveAll(validUserPasswordResetTokens);
-                        });
-                    } else if (token.getStatus().equals(TokenStatus.REVOKED)) {
-                        throw new VerificationException(TOKEN_REVOKED);
-                    } else if (token.getStatus().equals(TokenStatus.EXPIRED)) {
-                        throw new VerificationException(TOKEN_EXPIRED);
-                    } else {
-                        throw new VerificationException(JWT_TOKEN_INVALID);
-                    }
-                });
-    }
-
-    @Override
-    public void changeEmail(String email) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public AuthenticationResponse refreshToken(String refreshToken) {
-        String userEmail = jwtService.extractUsername(refreshToken);
-        if (userEmail == null) {
-            throw new VerificationException(JWT_TOKEN_INVALID);
-        }
-        var user = this.userRepository.getUserByEmailEquals(userEmail)
-                .orElseThrow(() -> new EntityNotFoundException(USER_NOT_FOUND));
-        if (jwtService.isTokenValid(refreshToken, user)) {
-            return getAuthenticationResponse(user);
-        }
-        throw new VerificationException(JWT_TOKEN_INVALID);
-    }
-
-    private AuthenticationResponse getAuthenticationResponse(User user) {
-        var accessToken = jwtService.generateAccessToken(user);
-        var refreshToken = jwtService.generateRefreshToken(user);
-        CompletableFuture.runAsync(() -> revokeAllUserAuthTokens(user));
-        saveUserToken(user, accessToken, TokenType.ACCESS, accessTokenExpiration);
-        saveUserToken(user, refreshToken, TokenType.REFRESH, refreshExpiration);
-        return AuthenticationResponse.builder()
-                .accessToken(accessToken)
-                .refreshToken(refreshToken)
-                .build();
-    }
-
-    private void revokeAllUserAuthTokens(User user) {
-        var validUserTokens = tokenRepository.findAllValidTokenByUser(user.getUserId())
-                .stream()
-                .filter(t -> t.getTokenType().equals(TokenType.ACCESS) || t.getTokenType().equals(TokenType.REFRESH))
-                .toList();
-        if (!validUserTokens.isEmpty()) {
-            validUserTokens.forEach(token -> token.setStatus(TokenStatus.REVOKED));
-            tokenRepository.saveAll(validUserTokens);
-        }
-    }
-
-    @Override
-    public void verifyForgotPasswordToken(String token) {
-        getUserFromToken(token);
-    }
-
-    @Override
-    public void verifyEmailToken(String token) {
-        var user = getUserFromToken(token);
-        CompletableFuture.runAsync(() -> {
-            var validUserEmailVerificationTokens = tokenRepository.findAllValidTokenByUser(user.getUserId())
-                    .stream()
-                    .filter(t -> t.getTokenType().equals(TokenType.EMAIL_VERIFICATION))
-                    .toList();
-            validUserEmailVerificationTokens.forEach(t -> t.setStatus(TokenStatus.REVOKED));
-            tokenRepository.saveAll(validUserEmailVerificationTokens);
-        });
-        user.setEmailVerified(true);
-        userRepository.save(user);
-    }
-
-    @Override
-    public User getUserFromToken(String token) {
-        String userEmail = jwtService.extractUsername(token);
-        if (userEmail == null) {
-            throw new VerificationException(VERIFICATION_TOKEN_INVALID);
-        }
-        var user = this.userRepository.getUserByEmailEquals(userEmail)
-                .orElseThrow(() -> new EntityNotFoundException(USER_NOT_FOUND));
-        if (!jwtService.isTokenValid(token, user)) {
-            throw new VerificationException(VERIFICATION_TOKEN_INVALID);
-        }
-        return user;
-    }
-}
