+++ src/main/resources/application-dev.yml
@@ -13,8 +13,6 @@ spring:
         use_nationalized_character_data: true
   rabbitmq:
     addresses: ${RABBITMQ_HOST:localhost}:5672
-    stomp:
-      port: 61613
   mail:
     host: smtp.gmail.com
     port: 587
