+++ src/main/resources/application.yml
@@ -19,8 +19,6 @@ spring:
     show-sql: true
   rabbitmq:
     addresses: ${RABBITMQ_HOST:localhost}:5672
-    stomp:
-      port: 61613
   mail:
     host: smtp.gmail.com
     port: 587
@@ -50,12 +48,7 @@ rabbitmq:
 mail:
   debug: true
 
-redis:
-  host: ${HOST:localhost}
-  port: 6379
-
 application:
-  host: ${HOST:localhost}
   base-url: ${BASE_URL:http://localhost:8080}
   security:
     jwt:
